feat(renderer): migrate index buffers from uint16 to uint32
GeometryData now accepts uint32_t indices and creates a kUInt32 index
buffer, lifting the 65535-vertex cap. MeshLoader's narrowing cast and
vertex-count guard are removed; all procedural geometry helpers
(GeometryUtils, FoliageRenderer, TreeRenderer) updated accordingly.

Close #420

+++ src/renderer/FoliageRenderer.cpp
@@ -65,7 +65,7 @@ std::unique_ptr<GeometryData> FoliageRenderer::MakeBillboardQuad() const {
     { { 0.5f, 1.f, 0.f}, n, b, t, {1.f, 0.f} },
     { {-0.5f, 1.f, 0.f}, n, b, t, {0.f, 0.f} },
   };
-  const uint16_t idx[6] = { 0, 1, 2, 0, 2, 3 };
+  const uint32_t idx[6] = { 0, 1, 2, 0, 2, 3 };
   return std::make_unique<GeometryData>(video_, 4, verts, 2, idx);
 }
